fix(audit): keep emitter below file-size budget
